Create class model App\Models\ListElement
Also add new table 'list_elements' to the database, and make dozens of
changes to the other model classes and existing unit tests: The 'elements'
array that was a simple property on the TaskList class was expanded to
become its own model class, with a one-to-many relationship to TaskList.
Methods for adding and deleting a TaskList element were shuffled over and
heavily modified to become static methods of the ListElement class.