Add form view-partials for task items to 'list.show' view
Also fix a bug I discovered while working on the view partials:
Task::updateDetails method did not check if a DeadlineItem was missing on
the Task before trying to update it. So a conditional was added that
checks if a DeadlineItem already exists. If so, it's updated as before.
Otherwise, a new one is created.