1st-draft implementation: User timezone saved when TaskList loads
Create new migration to add 'timezone' column to 'users' table. Add hidden
form to 'list.show' view that 1) holds value of user's previous timezone
offset in a data attribute, 2) gets value of user's current timezone
offset from a JS <script> below the form, 3) POSTs the timezone offset
to route 'user.timezone' via JS command if the offset has changed. Create
a new web-route for 'user.timezone' with a simple closure handler that
converts the offset to a timezone name, saves it to the User model, and
redirects back to the 'list.show' view.