Create taskList/StoreTimezoneForm.js. Activate via taskList/TaskList.js.
Move JS code from the (temporary) <script> element in the 'list.show' view to
the new StoreTimezoneForm class. Add a 'storeTimezoneForm' property to the
TaskList class, which instantiates StoreTimezoneForm. Call
storeTimezoneForm.activate() withing the TaskList.activate() method, which
gets the user's browser's timezone offset, compares it to the user's
stored offset, and submits the #storeTimezone form if they're different.